Split rerouting from TransportReplicationAction
Today's `TransportReplicationAction` routes the request to the primary
(following relocations etc) and then performs the action and replicates it. In
some cases we want to avoid the rerouting phases because the action is being
coordinated by the primary itself. This change separates out a
`TransportReroutedReplicationAction` which performs the rerouting and primary
delegation phases before handing over to `TransportReplicationAction` to do the
actual replication.
