+++ server/src/main/java/org/elasticsearch/action/resync/TransportResyncReplicationAction.java
@@ -85,8 +85,10 @@ public class TransportResyncReplicationAction extends TransportWriteAction<Resyn
     }
 
     @Override
-    protected void shardOperationOnPrimary(ResyncReplicationRequest request, IndexShard primary,
-            ActionListener<PrimaryResult<ResyncReplicationRequest, ResyncReplicationResponse>> listener) {
+    protected void shardOperationOnPrimary(
+        ResyncReplicationRequest request, IndexShard primary,
+        ActionListener<TransportReplicationAction.PrimaryResult<ResyncReplicationRequest, ResyncReplicationResponse>> listener) {
+
         ActionListener.completeWith(listener,
             () -> new WritePrimaryResult<>(performOnPrimary(request), new ResyncReplicationResponse(), null, null, primary, logger));
     }
