+++ server/src/main/java/org/elasticsearch/action/support/replication/TransportReroutedReplicationAction.java
@@ -1,0 +1,883 @@
+/*
+ * Licensed to Elasticsearch under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.elasticsearch.action.support.replication;
+
+import org.apache.logging.log4j.message.ParameterizedMessage;
+import org.elasticsearch.Assertions;
+import org.elasticsearch.action.ActionListener;
+import org.elasticsearch.action.ActionListenerResponseHandler;
+import org.elasticsearch.action.ActionRequest;
+import org.elasticsearch.action.ActionRequestValidationException;
+import org.elasticsearch.action.UnavailableShardsException;
+import org.elasticsearch.action.support.ActionFilters;
+import org.elasticsearch.action.support.ActiveShardCount;
+import org.elasticsearch.action.support.ChannelActionListener;
+import org.elasticsearch.action.support.TransportAction;
+import org.elasticsearch.action.support.TransportActions;
+import org.elasticsearch.cluster.ClusterState;
+import org.elasticsearch.cluster.ClusterStateObserver;
+import org.elasticsearch.cluster.action.shard.ShardStateAction;
+import org.elasticsearch.cluster.block.ClusterBlockException;
+import org.elasticsearch.cluster.block.ClusterBlockLevel;
+import org.elasticsearch.cluster.metadata.IndexMetaData;
+import org.elasticsearch.cluster.metadata.IndexNameExpressionResolver;
+import org.elasticsearch.cluster.node.DiscoveryNode;
+import org.elasticsearch.cluster.routing.AllocationId;
+import org.elasticsearch.cluster.routing.IndexShardRoutingTable;
+import org.elasticsearch.cluster.routing.ShardRouting;
+import org.elasticsearch.cluster.service.ClusterService;
+import org.elasticsearch.common.Nullable;
+import org.elasticsearch.common.io.stream.StreamInput;
+import org.elasticsearch.common.io.stream.StreamOutput;
+import org.elasticsearch.common.io.stream.Writeable;
+import org.elasticsearch.common.lease.Releasable;
+import org.elasticsearch.common.lease.Releasables;
+import org.elasticsearch.common.settings.Settings;
+import org.elasticsearch.common.unit.TimeValue;
+import org.elasticsearch.common.util.concurrent.AbstractRunnable;
+import org.elasticsearch.index.IndexNotFoundException;
+import org.elasticsearch.index.IndexService;
+import org.elasticsearch.index.shard.IndexShard;
+import org.elasticsearch.index.shard.ReplicationGroup;
+import org.elasticsearch.index.shard.ShardId;
+import org.elasticsearch.index.shard.ShardNotFoundException;
+import org.elasticsearch.indices.IndexClosedException;
+import org.elasticsearch.indices.IndicesService;
+import org.elasticsearch.node.NodeClosedException;
+import org.elasticsearch.tasks.Task;
+import org.elasticsearch.tasks.TaskId;
+import org.elasticsearch.threadpool.ThreadPool;
+import org.elasticsearch.transport.ConnectTransportException;
+import org.elasticsearch.transport.TransportChannel;
+import org.elasticsearch.transport.TransportException;
+import org.elasticsearch.transport.TransportRequest;
+import org.elasticsearch.transport.TransportRequestOptions;
+import org.elasticsearch.transport.TransportResponseHandler;
+import org.elasticsearch.transport.TransportService;
+
+import java.io.IOException;
+import java.util.Map;
+import java.util.Objects;
+import java.util.concurrent.atomic.AtomicBoolean;
+
+/**
+ * Base class for requests that should be routed to a primary shard copy, executed there, and then executed on all the replica copies.
+ * Subclasses can provide implementations for primary and replica operations.
+ *
+ * The action samples the cluster state on the receiving node to reroute to the node with the primary copy and on the
+ * primary node to validate the request before the primary operation followed by sampling the state again for resolving
+ * the nodes with replica copies to perform replication.
+ */
+public abstract class TransportReroutedReplicationAction<
+            Request extends ReplicationRequest<Request>,
+            ReplicaRequest extends ReplicationRequest<ReplicaRequest>,
+            Response extends ReplicationResponse
+        > extends TransportAction<Request, Response> {
+
+    protected final ThreadPool threadPool;
+    protected final TransportService transportService;
+    protected final ClusterService clusterService;
+    protected final ShardStateAction shardStateAction;
+    protected final IndicesService indicesService;
+    protected final IndexNameExpressionResolver indexNameExpressionResolver;
+    protected final TransportRequestOptions transportOptions;
+    protected final String executor;
+    protected final TransportReplicationAction<Request, ReplicaRequest, Response> transportReplicationAction;
+
+    // package private for testing
+    protected final String transportReplicaAction;
+    protected final String transportPrimaryAction;
+
+    protected TransportReroutedReplicationAction(Settings settings, String actionName, TransportService transportService,
+                                                 ClusterService clusterService, IndicesService indicesService,
+                                                 ThreadPool threadPool, ShardStateAction shardStateAction,
+                                                 ActionFilters actionFilters,
+                                                 IndexNameExpressionResolver indexNameExpressionResolver,
+                                                 Writeable.Reader<Request> requestReader,
+                                                 Writeable.Reader<ReplicaRequest> replicaRequestReader, String executor) {
+        this(settings, actionName, transportService, clusterService, indicesService, threadPool, shardStateAction, actionFilters,
+                indexNameExpressionResolver, requestReader, replicaRequestReader, executor, false, false);
+    }
+
+
+    protected TransportReroutedReplicationAction(Settings settings, String actionName, TransportService transportService,
+                                                 ClusterService clusterService, IndicesService indicesService,
+                                                 ThreadPool threadPool, ShardStateAction shardStateAction,
+                                                 ActionFilters actionFilters,
+                                                 IndexNameExpressionResolver indexNameExpressionResolver,
+                                                 Writeable.Reader<Request> requestReader,
+                                                 Writeable.Reader<ReplicaRequest> replicaRequestReader, String executor,
+                                                 boolean syncGlobalCheckpointAfterOperation, boolean forceExecutionOnPrimary) {
+        super(actionName, actionFilters, transportService.getTaskManager());
+        this.threadPool = threadPool;
+        this.transportService = transportService;
+        this.clusterService = clusterService;
+        this.indicesService = indicesService;
+        this.shardStateAction = shardStateAction;
+        this.indexNameExpressionResolver = indexNameExpressionResolver;
+        this.executor = executor;
+
+        this.transportPrimaryAction = actionName + "[p]";
+        this.transportReplicaAction = actionName + "[r]";
+
+        transportService.registerRequestHandler(actionName, ThreadPool.Names.SAME, requestReader, this::handleOperationRequest);
+
+        transportService.registerRequestHandler(transportPrimaryAction, executor, forceExecutionOnPrimary, true,
+            in -> new ConcreteShardRequest<>(requestReader, in), this::handlePrimaryRequest);
+
+        this.transportOptions = transportOptions(settings);
+
+        this.transportReplicationAction = new TransportReplicationAction<>
+            (actionName, actionFilters, taskManager, transportService, executor, replicaRequestReader, clusterService, threadPool,
+                indicesService, transportOptions, syncGlobalCheckpointAfterOperation) {
+
+            @Override
+            protected void acquireReplicaOperationPermit(IndexShard replica, ReplicaRequest request, ActionListener<Releasable> onAcquired,
+                                                         long primaryTerm, long globalCheckpoint, long maxSeqNoOfUpdatesOrDeletes) {
+
+                TransportReroutedReplicationAction.this.acquireReplicaOperationPermit(replica, request, onAcquired, primaryTerm,
+                    globalCheckpoint, maxSeqNoOfUpdatesOrDeletes);
+            }
+
+            @Override
+            protected ReplicaResult shardOperationOnReplica(ReplicaRequest shardRequest, IndexShard replica) throws Exception {
+                return TransportReroutedReplicationAction.this.shardOperationOnReplica(shardRequest, replica);
+            }
+        };
+    }
+
+    @Override
+    protected void doExecute(Task task, Request request, ActionListener<Response> listener) {
+        assert request.shardId() != null : "request shardId must be set";
+        new ReroutePhase((ReplicationTask) task, request, listener).run();
+    }
+
+    protected ReplicationOperation.Replicas<ReplicaRequest> newReplicasProxy() {
+        return new ReplicasProxy();
+    }
+
+    protected abstract Response newResponseInstance();
+
+    /**
+     * Resolves derived values in the request. For example, the target shard id of the incoming request, if not set at request construction.
+     * Additional processing or validation of the request should be done here.
+     *
+     * @param indexMetaData index metadata of the concrete index this request is going to operate on
+     * @param request       the request to resolve
+     */
+    protected void resolveRequest(final IndexMetaData indexMetaData, final Request request) {
+        if (request.waitForActiveShards() == ActiveShardCount.DEFAULT) {
+            // if the wait for active shard count has not been set in the request,
+            // resolve it from the index settings
+            request.waitForActiveShards(indexMetaData.getWaitForActiveShards());
+        }
+    }
+
+    /**
+     * Primary operation on node with primary copy.
+     *
+     * @param shardRequest the request to the primary shard
+     * @param primary      the primary shard to perform the operation on
+     */
+    protected abstract void shardOperationOnPrimary(Request shardRequest, IndexShard primary,
+        ActionListener<TransportReplicationAction.PrimaryResult<ReplicaRequest, Response>> listener);
+
+    /**
+     * Synchronously execute the specified replica operation. This is done under a permit from
+     * {@link IndexShard#acquireReplicaOperationPermit(long, long, long, ActionListener, String, Object)}.
+     *
+     * @param shardRequest the request to the replica shard
+     * @param replica      the replica shard to perform the operation on
+     */
+    protected abstract TransportReplicationAction.ReplicaResult
+        shardOperationOnReplica(ReplicaRequest shardRequest, IndexShard replica) throws Exception;
+
+    /**
+     * Cluster level block to check before request execution. Returning null means that no blocks need to be checked.
+     */
+    @Nullable
+    protected ClusterBlockLevel globalBlockLevel() {
+        return null;
+    }
+
+    /**
+     * Index level block to check before request execution. Returning null means that no blocks need to be checked.
+     */
+    @Nullable
+    public ClusterBlockLevel indexBlockLevel() {
+        return null;
+    }
+
+    /**
+     * True if provided index should be resolved when resolving request
+     */
+    protected boolean resolveIndex() {
+        return true;
+    }
+
+    protected TransportRequestOptions transportOptions(Settings settings) {
+        return TransportRequestOptions.EMPTY;
+    }
+
+    private String concreteIndex(final ClusterState state, final ReplicationRequest request) {
+        return resolveIndex() ? indexNameExpressionResolver.concreteSingleIndex(state, request).getName() : request.index();
+    }
+
+    private ClusterBlockException blockExceptions(final ClusterState state, final String indexName) {
+        ClusterBlockLevel globalBlockLevel = globalBlockLevel();
+        if (globalBlockLevel != null) {
+            ClusterBlockException blockException = state.blocks().globalBlockedException(globalBlockLevel);
+            if (blockException != null) {
+                return blockException;
+            }
+        }
+        ClusterBlockLevel indexBlockLevel = indexBlockLevel();
+        if (indexBlockLevel != null) {
+            ClusterBlockException blockException = state.blocks().indexBlockedException(indexBlockLevel, indexName);
+            if (blockException != null) {
+                return blockException;
+            }
+        }
+        return null;
+    }
+
+    protected boolean retryPrimaryException(final Throwable e) {
+        return e.getClass() == ReplicationOperation.RetryOnPrimaryException.class
+                || TransportActions.isShardNotAvailableException(e)
+                || isRetryableClusterBlockException(e);
+    }
+
+    boolean isRetryableClusterBlockException(final Throwable e) {
+        if (e instanceof ClusterBlockException) {
+            return ((ClusterBlockException) e).retryable();
+        }
+        return false;
+    }
+
+    protected void handleOperationRequest(final Request request, final TransportChannel channel, Task task) {
+        execute(task, request, new ChannelActionListener<>(channel, actionName, request));
+    }
+
+    protected void handlePrimaryRequest(final ConcreteShardRequest<Request> request, final TransportChannel channel, final Task task) {
+        new AsyncPrimaryAction(
+            request, new ChannelActionListener<>(channel, transportPrimaryAction, request), (ReplicationTask) task).run();
+    }
+
+    class AsyncPrimaryAction extends AbstractRunnable {
+        private final ActionListener<Response> onCompletionListener;
+        private final ReplicationTask replicationTask;
+        private final ConcreteShardRequest<Request> primaryRequest;
+
+        AsyncPrimaryAction(ConcreteShardRequest<Request> primaryRequest, ActionListener<Response> onCompletionListener,
+                           ReplicationTask replicationTask) {
+            this.primaryRequest = primaryRequest;
+            this.onCompletionListener = onCompletionListener;
+            this.replicationTask = replicationTask;
+        }
+
+        @Override
+        protected void doRun() throws Exception {
+            final ShardId shardId = primaryRequest.getRequest().shardId();
+            final IndexShard indexShard = getIndexShard(shardId);
+            final ShardRouting shardRouting = indexShard.routingEntry();
+            // we may end up here if the cluster state used to route the primary is so stale that the underlying
+            // index shard was replaced with a replica. For example - in a two node cluster, if the primary fails
+            // the replica will take over and a replica will be assigned to the first node.
+            if (shardRouting.primary() == false) {
+                throw new ReplicationOperation.RetryOnPrimaryException(shardId, "actual shard is not a primary " + shardRouting);
+            }
+            final String actualAllocationId = shardRouting.allocationId().getId();
+            if (actualAllocationId.equals(primaryRequest.getTargetAllocationID()) == false) {
+                throw new ShardNotFoundException(shardId, "expected allocation id [{}] but found [{}]",
+                    primaryRequest.getTargetAllocationID(), actualAllocationId);
+            }
+            final long actualTerm = indexShard.getPendingPrimaryTerm();
+            if (actualTerm != primaryRequest.getPrimaryTerm()) {
+                throw new ShardNotFoundException(shardId, "expected allocation id [{}] with term [{}] but found [{}]",
+                    primaryRequest.getTargetAllocationID(), primaryRequest.getPrimaryTerm(), actualTerm);
+            }
+
+            acquirePrimaryOperationPermit(indexShard, primaryRequest.getRequest(), ActionListener.wrap(
+                releasable -> runWithPrimaryShardReference(new PrimaryShardReference(indexShard, releasable)),
+                this::onFailure
+            ));
+        }
+
+        void runWithPrimaryShardReference(final PrimaryShardReference primaryShardReference) {
+            try {
+                final ClusterState clusterState = clusterService.state();
+                final IndexMetaData indexMetaData = clusterState.metaData().getIndexSafe(primaryShardReference.routingEntry().index());
+
+                final ClusterBlockException blockException = blockExceptions(clusterState, indexMetaData.getIndex().getName());
+                if (blockException != null) {
+                    logger.trace("cluster is blocked, action failed on primary", blockException);
+                    throw blockException;
+                }
+
+                if (primaryShardReference.isRelocated()) {
+                    primaryShardReference.close(); // release shard operation lock as soon as possible
+                    setPhase(replicationTask, "primary_delegation");
+                    // delegate primary phase to relocation target
+                    // it is safe to execute primary phase on relocation target as there are no more in-flight operations where primary
+                    // phase is executed on local shard and all subsequent operations are executed on relocation target as primary phase.
+                    final ShardRouting primary = primaryShardReference.routingEntry();
+                    assert primary.relocating() : "indexShard is marked as relocated but routing isn't" + primary;
+                    final Writeable.Reader<Response> reader = in -> {
+                        Response response = TransportReroutedReplicationAction.this.newResponseInstance();
+                        response.readFrom(in);
+                        return response;
+                    };
+                    DiscoveryNode relocatingNode = clusterState.nodes().get(primary.relocatingNodeId());
+                    transportService.sendRequest(relocatingNode, transportPrimaryAction,
+                        new ConcreteShardRequest<>(primaryRequest.getRequest(), primary.allocationId().getRelocationId(),
+                            primaryRequest.getPrimaryTerm()),
+                        transportOptions,
+                        new ActionListenerResponseHandler<>(onCompletionListener, reader) {
+                            @Override
+                            public void handleResponse(Response response) {
+                                setPhase(replicationTask, "finished");
+                                super.handleResponse(response);
+                            }
+
+                            @Override
+                            public void handleException(TransportException exp) {
+                                setPhase(replicationTask, "finished");
+                                super.handleException(exp);
+                            }
+                        });
+                } else {
+                    setPhase(replicationTask, "primary");
+
+                    final ActionListener<Response> referenceClosingListener = ActionListener.wrap(response -> {
+                        primaryShardReference.close(); // release shard operation lock before responding to caller
+                        setPhase(replicationTask, "finished");
+                        onCompletionListener.onResponse(response);
+                    }, e -> handleException(primaryShardReference, e));
+
+                    assert primaryShardReference.indexShard.getOperationPrimaryTerm() == primaryRequest.getPrimaryTerm()
+                        : primaryShardReference.indexShard.getOperationPrimaryTerm() + " vs " + primaryRequest.getPrimaryTerm();
+
+                    assert primaryShardReference.routingEntry().allocationId().getId().equals(primaryRequest.getTargetAllocationID())
+                        : primaryShardReference.routingEntry().allocationId().getId() + " vs " + primaryRequest.getTargetAllocationID();
+
+                    transportReplicationAction.performAndReplicate(primaryRequest.getRequest(), primaryShardReference.indexShard,
+                        primaryShardReference, newReplicasProxy(), referenceClosingListener);
+                }
+            } catch (Exception e) {
+                handleException(primaryShardReference, e);
+            }
+        }
+
+        private void handleException(PrimaryShardReference primaryShardReference, Exception e) {
+            Releasables.closeWhileHandlingException(primaryShardReference); // release shard operation lock before responding to caller
+            onFailure(e);
+        }
+
+        @Override
+        public void onFailure(Exception e) {
+            setPhase(replicationTask, "finished");
+            onCompletionListener.onFailure(e);
+        }
+    }
+
+    // for testing
+    void handleReplicaRequest(final TransportReplicationAction.ConcreteReplicaRequest<ReplicaRequest> replicaRequest,
+                              final ActionListener<TransportReplicationAction.ReplicaResponse> listener,
+                              final Task task) {
+        transportReplicationAction.execute(task, replicaRequest, listener);
+    }
+
+    private IndexShard getIndexShard(final ShardId shardId) {
+        IndexService indexService = indicesService.indexServiceSafe(shardId.getIndex());
+        return indexService.getShard(shardId.id());
+    }
+
+    /**
+     * Responsible for routing and retrying failed operations on the primary.
+     * The actual primary operation is done in {@link ReplicationOperation} on the
+     * node with primary copy.
+     *
+     * Resolves index and shard id for the request before routing it to target node
+     */
+    final class ReroutePhase extends AbstractRunnable {
+        private final ActionListener<Response> listener;
+        private final Request request;
+        private final ReplicationTask task;
+        private final ClusterStateObserver observer;
+        private final AtomicBoolean finished = new AtomicBoolean();
+
+        ReroutePhase(ReplicationTask task, Request request, ActionListener<Response> listener) {
+            this.request = request;
+            if (task != null) {
+                this.request.setParentTask(clusterService.localNode().getId(), task.getId());
+            }
+            this.listener = listener;
+            this.task = task;
+            this.observer = new ClusterStateObserver(clusterService, request.timeout(), logger, threadPool.getThreadContext());
+        }
+
+        @Override
+        public void onFailure(Exception e) {
+            finishWithUnexpectedFailure(e);
+        }
+
+        @Override
+        protected void doRun() {
+            setPhase(task, "routing");
+            final ClusterState state = observer.setAndGetObservedState();
+            final String concreteIndex = concreteIndex(state, request);
+            final ClusterBlockException blockException = blockExceptions(state, concreteIndex);
+            if (blockException != null) {
+                if (blockException.retryable()) {
+                    logger.trace("cluster is blocked, scheduling a retry", blockException);
+                    retry(blockException);
+                } else {
+                    finishAsFailed(blockException);
+                }
+            } else {
+                // request does not have a shardId yet, we need to pass the concrete index to resolve shardId
+                final IndexMetaData indexMetaData = state.metaData().index(concreteIndex);
+                if (indexMetaData == null) {
+                    retry(new IndexNotFoundException(concreteIndex));
+                    return;
+                }
+                if (indexMetaData.getState() == IndexMetaData.State.CLOSE) {
+                    throw new IndexClosedException(indexMetaData.getIndex());
+                }
+
+                // resolve all derived request fields, so we can route and apply it
+                resolveRequest(indexMetaData, request);
+                assert request.waitForActiveShards() != ActiveShardCount.DEFAULT :
+                    "request waitForActiveShards must be set in resolveRequest";
+
+                final ShardRouting primary = primary(state);
+                if (retryIfUnavailable(state, primary)) {
+                    return;
+                }
+                final DiscoveryNode node = state.nodes().get(primary.currentNodeId());
+                if (primary.currentNodeId().equals(state.nodes().getLocalNodeId())) {
+                    performLocalAction(state, primary, node, indexMetaData);
+                } else {
+                    performRemoteAction(state, primary, node);
+                }
+            }
+        }
+
+        private void performLocalAction(ClusterState state, ShardRouting primary, DiscoveryNode node, IndexMetaData indexMetaData) {
+            setPhase(task, "waiting_on_primary");
+            if (logger.isTraceEnabled()) {
+                logger.trace("send action [{}] to local primary [{}] for request [{}] with cluster state version [{}] to [{}] ",
+                    transportPrimaryAction, request.shardId(), request, state.version(), primary.currentNodeId());
+            }
+            performAction(node, transportPrimaryAction, true,
+                new ConcreteShardRequest<>(request, primary.allocationId().getId(), indexMetaData.primaryTerm(primary.id())));
+        }
+
+        private void performRemoteAction(ClusterState state, ShardRouting primary, DiscoveryNode node) {
+            if (state.version() < request.routedBasedOnClusterVersion()) {
+                logger.trace("failed to find primary [{}] for request [{}] despite sender thinking it would be here. Local cluster state "
+                        + "version [{}]] is older than on sending node (version [{}]), scheduling a retry...", request.shardId(), request,
+                    state.version(), request.routedBasedOnClusterVersion());
+                retryBecauseUnavailable(request.shardId(), "failed to find primary as current cluster state with version ["
+                    + state.version() + "] is stale (expected at least [" + request.routedBasedOnClusterVersion() + "]");
+                return;
+            } else {
+                // chasing the node with the active primary for a second hop requires that we are at least up-to-date with the current
+                // cluster state version this prevents redirect loops between two nodes when a primary was relocated and the relocation
+                // target is not aware that it is the active primary shard already.
+                request.routedBasedOnClusterVersion(state.version());
+            }
+            if (logger.isTraceEnabled()) {
+                logger.trace("send action [{}] on primary [{}] for request [{}] with cluster state version [{}] to [{}]", actionName,
+                    request.shardId(), request, state.version(), primary.currentNodeId());
+            }
+            setPhase(task, "rerouted");
+            performAction(node, actionName, false, request);
+        }
+
+        private boolean retryIfUnavailable(ClusterState state, ShardRouting primary) {
+            if (primary == null || primary.active() == false) {
+                logger.trace("primary shard [{}] is not yet active, scheduling a retry: action [{}], request [{}], "
+                    + "cluster state version [{}]", request.shardId(), actionName, request, state.version());
+                retryBecauseUnavailable(request.shardId(), "primary shard is not active");
+                return true;
+            }
+            if (state.nodes().nodeExists(primary.currentNodeId()) == false) {
+                logger.trace("primary shard [{}] is assigned to an unknown node [{}], scheduling a retry: action [{}], request [{}], "
+                    + "cluster state version [{}]", request.shardId(), primary.currentNodeId(), actionName, request, state.version());
+                retryBecauseUnavailable(request.shardId(), "primary shard isn't assigned to a known node.");
+                return true;
+            }
+            return false;
+        }
+
+        private ShardRouting primary(ClusterState state) {
+            IndexShardRoutingTable indexShard = state.getRoutingTable().shardRoutingTable(request.shardId());
+            return indexShard.primaryShard();
+        }
+
+        private void performAction(final DiscoveryNode node, final String action, final boolean isPrimaryAction,
+                                   final TransportRequest requestToPerform) {
+            transportService.sendRequest(node, action, requestToPerform, transportOptions, new TransportResponseHandler<Response>() {
+
+                @Override
+                public Response read(StreamInput in) throws IOException {
+                    Response response = newResponseInstance();
+                    response.readFrom(in);
+                    return response;
+                }
+
+                @Override
+                public String executor() {
+                    return ThreadPool.Names.SAME;
+                }
+
+                @Override
+                public void handleResponse(Response response) {
+                    finishOnSuccess(response);
+                }
+
+                @Override
+                public void handleException(TransportException exp) {
+                    try {
+                        // if we got disconnected from the node, or the node / shard is not in the right state (being closed)
+                        final Throwable cause = exp.unwrapCause();
+                        if (cause instanceof ConnectTransportException || cause instanceof NodeClosedException ||
+                            (isPrimaryAction && retryPrimaryException(cause))) {
+                            logger.trace(() -> new ParameterizedMessage(
+                                    "received an error from node [{}] for request [{}], scheduling a retry",
+                                    node.getId(), requestToPerform), exp);
+                            retry(exp);
+                        } else {
+                            finishAsFailed(exp);
+                        }
+                    } catch (Exception e) {
+                        e.addSuppressed(exp);
+                        finishWithUnexpectedFailure(e);
+                    }
+                }
+            });
+        }
+
+        void retry(Exception failure) {
+            assert failure != null;
+            if (observer.isTimedOut()) {
+                // we running as a last attempt after a timeout has happened. don't retry
+                finishAsFailed(failure);
+                return;
+            }
+            setPhase(task, "waiting_for_retry");
+            request.onRetry();
+            observer.waitForNextChange(new ClusterStateObserver.Listener() {
+                @Override
+                public void onNewClusterState(ClusterState state) {
+                    run();
+                }
+
+                @Override
+                public void onClusterServiceClose() {
+                    finishAsFailed(new NodeClosedException(clusterService.localNode()));
+                }
+
+                @Override
+                public void onTimeout(TimeValue timeout) {
+                    // Try one more time...
+                    run();
+                }
+            });
+        }
+
+        void finishAsFailed(Exception failure) {
+            if (finished.compareAndSet(false, true)) {
+                setPhase(task, "failed");
+                logger.trace(() -> new ParameterizedMessage("operation failed. action [{}], request [{}]", actionName, request), failure);
+                listener.onFailure(failure);
+            } else {
+                assert false : "finishAsFailed called but operation is already finished";
+            }
+        }
+
+        void finishWithUnexpectedFailure(Exception failure) {
+            logger.warn(() -> new ParameterizedMessage(
+                        "unexpected error during the primary phase for action [{}], request [{}]",
+                        actionName, request), failure);
+            if (finished.compareAndSet(false, true)) {
+                setPhase(task, "failed");
+                listener.onFailure(failure);
+            } else {
+                assert false : "finishWithUnexpectedFailure called but operation is already finished";
+            }
+        }
+
+        void finishOnSuccess(Response response) {
+            if (finished.compareAndSet(false, true)) {
+                setPhase(task, "finished");
+                if (logger.isTraceEnabled()) {
+                    logger.trace("operation succeeded. action [{}],request [{}]", actionName, request);
+                }
+                listener.onResponse(response);
+            } else {
+                assert false : "finishOnSuccess called but operation is already finished";
+            }
+        }
+
+        void retryBecauseUnavailable(ShardId shardId, String message) {
+            retry(new UnavailableShardsException(shardId, "{} Timeout: [{}], request: [{}]", message, request.timeout(), request));
+        }
+    }
+
+    /**
+     * Executes the logic for acquiring one or more operation permit on a primary shard. The default is to acquire a single permit but this
+     * method can be overridden to acquire more.
+     */
+    protected void acquirePrimaryOperationPermit(final IndexShard primary,
+                                                 final Request request,
+                                                 final ActionListener<Releasable> onAcquired) {
+        primary.acquirePrimaryOperationPermit(onAcquired, executor, request);
+    }
+
+    /**
+     * Executes the logic for acquiring one or more operation permit on a replica shard. The default is to acquire a single permit but this
+     * method can be overridden to acquire more.
+     */
+    protected void acquireReplicaOperationPermit(final IndexShard replica,
+                                                 final ReplicaRequest request,
+                                                 final ActionListener<Releasable> onAcquired,
+                                                 final long primaryTerm,
+                                                 final long globalCheckpoint,
+                                                 final long maxSeqNoOfUpdatesOrDeletes) {
+        replica.acquireReplicaOperationPermit(primaryTerm, globalCheckpoint, maxSeqNoOfUpdatesOrDeletes, onAcquired, executor, request);
+    }
+
+    class PrimaryShardReference implements Releasable, ReplicationOperation.Primary<
+        Request, ReplicaRequest, TransportReplicationAction.PrimaryResult<ReplicaRequest, Response>> {
+
+        protected final IndexShard indexShard;
+        private final Releasable operationLock;
+
+        PrimaryShardReference(IndexShard indexShard, Releasable operationLock) {
+            this.indexShard = indexShard;
+            this.operationLock = operationLock;
+        }
+
+        @Override
+        public void close() {
+            operationLock.close();
+        }
+
+        public long getLocalCheckpoint() {
+            return indexShard.getLocalCheckpoint();
+        }
+
+        public ShardRouting routingEntry() {
+            return indexShard.routingEntry();
+        }
+
+        public boolean isRelocated() {
+            return indexShard.isRelocatedPrimary();
+        }
+
+        @Override
+        public void failShard(String reason, Exception e) {
+            try {
+                indexShard.failShard(reason, e);
+            } catch (Exception inner) {
+                e.addSuppressed(inner);
+            }
+        }
+
+        @Override
+        public void perform(Request request,
+                            ActionListener<TransportReplicationAction.PrimaryResult<ReplicaRequest, Response>> listener) {
+            if (Assertions.ENABLED) {
+                listener = ActionListener.map(listener, result -> {
+                    assert result.replicaRequest() == null || result.finalFailure == null : "a replica request [" + result.replicaRequest()
+                        + "] with a primary failure [" + result.finalFailure + "]";
+                    return result;
+                });
+            }
+            assert indexShard.getActiveOperationsCount() != 0 : "must perform shard operation under a permit";
+            shardOperationOnPrimary(request, indexShard, listener);
+        }
+
+        @Override
+        public void updateLocalCheckpointForShard(String allocationId, long checkpoint) {
+            indexShard.updateLocalCheckpointForShard(allocationId, checkpoint);
+        }
+
+        @Override
+        public void updateGlobalCheckpointForShard(final String allocationId, final long globalCheckpoint) {
+            indexShard.updateGlobalCheckpointForShard(allocationId, globalCheckpoint);
+        }
+
+        @Override
+        public long localCheckpoint() {
+            return indexShard.getLocalCheckpoint();
+        }
+
+        @Override
+        public long globalCheckpoint() {
+            return indexShard.getGlobalCheckpoint();
+        }
+
+        @Override
+        public long maxSeqNoOfUpdatesOrDeletes() {
+            return indexShard.getMaxSeqNoOfUpdatesOrDeletes();
+        }
+
+        @Override
+        public ReplicationGroup getReplicationGroup() {
+            return indexShard.getReplicationGroup();
+        }
+    }
+
+
+    /**
+     * The {@code ReplicasProxy} is an implementation of the {@code Replicas}
+     * interface that performs the actual {@code ReplicaRequest} on the replica
+     * shards. It also encapsulates the logic required for failing the replica
+     * if deemed necessary as well as marking it as stale when needed.
+     */
+    protected class ReplicasProxy implements ReplicationOperation.Replicas<ReplicaRequest> {
+
+        @Override
+        public void performOn(
+                final ShardRouting replica,
+                final ReplicaRequest request,
+                final long primaryTerm,
+                final long globalCheckpoint,
+                final long maxSeqNoOfUpdatesOrDeletes,
+                final ActionListener<ReplicationOperation.ReplicaResponse> listener) {
+            transportReplicationAction.performOnReplica(
+                replica, request, primaryTerm, globalCheckpoint, maxSeqNoOfUpdatesOrDeletes, listener);
+        }
+
+        @Override
+        public void failShardIfNeeded(ShardRouting replica, long primaryTerm, String message, Exception exception,
+                                      ActionListener<Void> listener) {
+            // This does not need to fail the shard. The idea is that this
+            // is a non-write operation (something like a refresh or a global
+            // checkpoint sync) and therefore the replica should still be
+            // "alive" if it were to fail.
+            listener.onResponse(null);
+        }
+
+        @Override
+        public void markShardCopyAsStaleIfNeeded(ShardId shardId, String allocationId, long primaryTerm, ActionListener<Void> listener) {
+            // This does not need to make the shard stale. The idea is that this
+            // is a non-write operation (something like a refresh or a global
+            // checkpoint sync) and therefore the replica should still be
+            // "alive" if it were to be marked as stale.
+            listener.onResponse(null);
+        }
+    }
+
+    /** a wrapper class to encapsulate a request when being sent to a specific allocation id **/
+    public static class ConcreteShardRequest<R extends TransportRequest> extends ActionRequest {
+
+        /** {@link AllocationId#getId()} of the shard this request is sent to **/
+        private final String targetAllocationID;
+        private final long primaryTerm;
+        private final R request;
+
+        public ConcreteShardRequest(Writeable.Reader<R> requestReader, StreamInput in) throws IOException {
+            targetAllocationID = in.readString();
+            primaryTerm  = in.readVLong();
+            request = requestReader.read(in);
+        }
+
+        public ConcreteShardRequest(R request, String targetAllocationID, long primaryTerm) {
+            Objects.requireNonNull(request);
+            Objects.requireNonNull(targetAllocationID);
+            this.request = request;
+            this.targetAllocationID = targetAllocationID;
+            this.primaryTerm = primaryTerm;
+        }
+
+        @Override
+        public void setParentTask(String parentTaskNode, long parentTaskId) {
+            request.setParentTask(parentTaskNode, parentTaskId);
+        }
+
+        @Override
+        public void setParentTask(TaskId taskId) {
+            request.setParentTask(taskId);
+        }
+
+        @Override
+        public TaskId getParentTask() {
+            return request.getParentTask();
+        }
+        @Override
+        public Task createTask(long id, String type, String action, TaskId parentTaskId, Map<String, String> headers) {
+            return request.createTask(id, type, action, parentTaskId, headers);
+        }
+
+        @Override
+        public String getDescription() {
+            return "[" + request.getDescription() + "] for aID [" + targetAllocationID + "] and term [" + primaryTerm + "]";
+        }
+
+        @Override
+        public ActionRequestValidationException validate() {
+            return null;
+        }
+
+        @Override
+        public void readFrom(StreamInput in) {
+            throw new UnsupportedOperationException("usage of Streamable is to be replaced by Writeable");
+        }
+
+        @Override
+        public void writeTo(StreamOutput out) throws IOException {
+            out.writeString(targetAllocationID);
+            out.writeVLong(primaryTerm);
+            request.writeTo(out);
+        }
+
+        public R getRequest() {
+            return request;
+        }
+
+        public String getTargetAllocationID() {
+            return targetAllocationID;
+        }
+
+        public long getPrimaryTerm() {
+            return primaryTerm;
+        }
+
+        @Override
+        public String toString() {
+            return "request: " + request + ", target allocation id: " + targetAllocationID + ", primary term: " + primaryTerm;
+        }
+    }
+
+    /**
+     * Sets the current phase on the task if it isn't null. Pulled into its own
+     * method because its more convenient that way.
+     */
+    private static void setPhase(ReplicationTask task, String phase) {
+        if (task != null) {
+            task.setPhase(phase);
+        }
+    }
+}
+
