+++ server/src/test/java/org/elasticsearch/action/IndicesRequestIT.java
@@ -70,7 +70,7 @@ import org.elasticsearch.action.search.SearchRequest;
 import org.elasticsearch.action.search.SearchResponse;
 import org.elasticsearch.action.search.SearchTransportService;
 import org.elasticsearch.action.search.SearchType;
-import org.elasticsearch.action.support.replication.TransportReplicationActionTests;
+import org.elasticsearch.action.support.replication.TransportReroutedReplicationActionTests;
 import org.elasticsearch.action.termvectors.MultiTermVectorsAction;
 import org.elasticsearch.action.termvectors.MultiTermVectorsRequest;
 import org.elasticsearch.action.termvectors.TermVectorsAction;
@@ -635,7 +635,7 @@ public class IndicesRequestIT extends ESIntegTestCase {
         if (request instanceof IndicesRequest) {
             indicesRequest = (IndicesRequest) request;
         } else {
-            indicesRequest = TransportReplicationActionTests.resolveRequest(request);
+            indicesRequest = TransportReroutedReplicationActionTests.resolveRequest(request);
         }
         return indicesRequest;
     }
